+++ modules/versions.lua
@@ -10,7 +10,7 @@ return {
   ["Gen1MenuManager"] = "0.5.0",
   ["Gen1ModMenu"] = "0.9.0",
   ["Gen1Rematch"] = "maintained",
-  ["Gen1Remember"] = "1.0.1",
+  ["Gen1Remember"] = "1.0.2",
   ["Gen1SoundQOL"] = "0.3.0",
   ["Gen1Sprint"] = "0.3.1",
   ["QualityOfLife"] = "maintained",